Custom indexing of Mathematica ".nb" files.
Specifically, this enables a special mode during indexing of Mathematica .nb files.

This mode only indexes the true contents of the notebook, skipping the metadata preamble and postambles which contain a bunch of junk.

It also keeps a running trigram to quickly recognize the location of CompressedData and CompressedBitmap structures, which contain randomly-distributed bytes that completely pollute the trigram index (or trigger the too-many-trigrams abort path).

We also skip indexing digits that occur in real numbers, because again they are usually randomly distributed, and often quite long.